# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SecureBank Core Banking API
  description: "Autenticação (JWT, refresh com rotação, MFA TOTP), clientes, contas,\
    \ depósitos, saques, transferências, pagamentos, extrato, limites, auditoria e\
    \ administração. Valores monetários trafegam como string (\"100.00\"). Transferê\
    ncias e pagamentos exigem o header Idempotency-Key."
  version: v1
servers:
- url: http://localhost:8100
  description: Generated server url
security:
- bearerAuth: []
paths:
  /api/v1/admin/accounts/{accountId}/limits/{type}:
    put:
      tags:
      - admin-controller
      operationId: changeLimit
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: string
      - name: type
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LimitRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LimitView"
...
  /api/v1/admin/users:
    post:
      tags:
      - admin-controller
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserView"
  /api/v1/admin/users/{id}/enable:
    post:
      tags:
      - admin-controller
      operationId: enableUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
  /api/v1/admin/users/{id}/disable:
    post:
      tags:
      - admin-controller
      operationId: disableUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
  /api/v1/admin/accounts/{id}/unblock:
    post:
      tags:
      - admin-controller
      operationId: unblock
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AccountStatusView"
  /api/v1/admin/accounts/{id}/block:
    post:
      tags:
      - admin-controller
      operationId: block
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AccountStatusView"
  /api/v1/accounts:
    get:
      tags:
      - account-controller
      operationId: list_2
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountResponse"
    post:
      tags:
      - account-controller
      operationId: open
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenAccountRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AccountResponse"
  /api/v1/accounts/{id}/withdrawals:
    post:
      tags:
      - account-controller
      operationId: withdraw
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmountRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransactionResponse"
  /api/v1/accounts/{id}/deposits:
    post:
      tags:
      - account-controller
      operationId: deposit
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmountRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransactionResponse"
...
  /api/v1/admin/customers/{id}:
    get:
      tags:
      - admin-controller
      operationId: customer
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CustomerView"
  /api/v1/accounts/{id}:
    get:
      tags:
      - account-controller
      operationId: get_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AccountResponse"
  /api/v1/accounts/{id}/statement:
    get:
      tags:
      - account-controller
      operationId: statement
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/StatementResponse"
  /api/v1/accounts/{id}/limits:
    get:
      tags:
      - account-controller
      operationId: limits
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LimitResponse"
  /api/v1/accounts/{id}/balance:
    get:
      tags:
      - account-controller
      operationId: balance
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BalanceResponse"
...
components:
  schemas:
    LimitRequest:
      type: object
      properties:
        perOperation:
          type: number
          minimum: 0.01
        daily:
          type: number
          minimum: 0.01
      required:
      - daily
      - perOperation
    LimitView:
      type: object
      properties:
        type:
          type: string
        perOperation:
          $ref: "#/components/schemas/MoneyResponse"
        daily:
          $ref: "#/components/schemas/MoneyResponse"
    MoneyResponse:
      type: object
      properties:
        amount:
          type: string
        currency:
          type: string
...
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          maxLength: 254
          minLength: 0
        password:
          type: string
          maxLength: 128
          minLength: 0
        role:
          type: string
          enum:
          - CUSTOMER
          - SUPPORT
          - ADMIN
      required:
      - email
      - password
      - role
    UserView:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
        status:
          type: string
    AccountStatusView:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
    OpenAccountRequest:
      type: object
      properties:
        type:
          type: string
          enum:
          - CHECKING
          - SAVINGS
      required:
      - type
    AccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        branch:
          type: string
        accountNumber:
          type: string
        type:
          type: string
        status:
          type: string
        balance:
          $ref: "#/components/schemas/MoneyResponse"
        createdAt:
          type: string
          format: date-time
    AmountRequest:
      type: object
      properties:
        amount:
          type: number
          minimum: 0.01
      required:
      - amount
    TransactionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        type:
          type: string
        direction:
          type: string
        amount:
          $ref: "#/components/schemas/MoneyResponse"
        balanceAfter:
          $ref: "#/components/schemas/MoneyResponse"
        status:
          type: string
        reference:
          type: string
        createdAt:
          type: string
          format: date-time
...
    CustomerView:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        document:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    StatementResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TransactionResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
    LimitResponse:
      type: object
      properties:
        type:
          type: string
        perOperation:
          $ref: "#/components/schemas/MoneyResponse"
        daily:
          $ref: "#/components/schemas/MoneyResponse"
        usedToday:
          $ref: "#/components/schemas/MoneyResponse"
        remainingToday:
          $ref: "#/components/schemas/MoneyResponse"
    BalanceResponse:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        balance:
          $ref: "#/components/schemas/MoneyResponse"
        asOf:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      description: "Access token (RS256, 15 min) obtido em POST /api/v1/auth/login"
      scheme: bearer
      bearerFormat: JWT